Added separate files for api, constants and utils
