+++ constants.js
@@ -1,0 +1,26 @@
+const CRYPTO_SYMBOLS = [
+  "BTC",
+  "ETH",
+  "BNB",
+  "SOL",
+  "XRP",
+  "TON",
+  "NOT",
+  "DOGS",
+];
+const STOCK_SYMBOLS = [
+  "AAPL",
+  "MSFT",
+  "AMZN",
+  "NVDA",
+  "GOOGL",
+  "TSLA",
+  "META",
+  "VOO",
+  "BZUN",
+  "BIDU",
+];
+
+const socialLinks = `🖼<a href="https://t.me/gunyainvest">Telegram</a> 📺<a href="https://www.youtube.com/@gunyainvest">YouTube</a> 💬<a href="https://t.me/investMoldova">Чат Invest Moldova</a> 💵<a href="https://patreon.com/GUNYAINVEST?utm_medium=unknown&utm_source=join_link&utm_campaign=creatorshare_creator&utm_content=copyLink">Patreon</a> 🎶<a href="https://www.tiktok.com/@investmoldova">TikTok</a>`;
+
+module.exports = { CRYPTO_SYMBOLS, STOCK_SYMBOLS, socialLinks };
